fix: fix circular dependency
