+++ src/legacy/components/Popups/FollowPendingTxPopup/FollowPendingTxPopupUI.tsx
@@ -4,7 +4,7 @@ import { Text } from 'rebass'
 import styled from 'styled-components/macro'
 
 import { AutoColumn } from 'legacy/components/Column'
-import { StyledClose as IconClose } from 'legacy/components/Popups/PopupItemMod'
+import { StyledClose as IconClose } from 'legacy/components/Popups/styled'
 import Tooltip, { TooltipProps } from 'legacy/components/Tooltip'
 
 interface PopupContentProps {
