+++ src/legacy/components/Popups/PopupItem.tsx
@@ -1,34 +1,15 @@
-import styled from 'styled-components/macro'
+import { useCallback, useContext, useEffect } from 'react'
 
+import { useSpring } from '@react-spring/web'
+import { ThemeContext } from 'styled-components/macro'
+
+import FailedNetworkSwitchPopup from 'legacy/components/Popups/FailedNetworkSwitchPopup'
+import { WarningPopup } from 'legacy/components/Popups/WarningPopup'
+import { useRemovePopup } from 'legacy/state/application/hooks'
 import { PopupContent } from 'legacy/state/application/reducer'
 
-import { default as PopupItemUni, Popup, Fader } from './PopupItemMod'
-
-const Wrapper = styled.div`
-  ${Popup} {
-    ${({ theme }) => theme.mediaWidth.upToSmall`
-      margin: 0 0 16px;
-      min-width: 100%;
-    `}
-  }
-
-  ${Fader} {
-    background-color: ${({ theme }) => theme.disabled};
-    height: 4px;
-  }
-
-  a {
-    text-decoration: underline;
-    color: ${({ theme }) => theme.textLink};
-  }
-`
-
-export const PopupItemWrapper = styled(PopupItemUni)`
-  ${(props) => props.className} {
-    border: 2px solid ${({ theme }) => theme.black};
-    box-shadow: 2px 2px 0 ${({ theme }) => theme.black};
-  }
-`
+import { AnimatedFader, PopupItemWrapper, PopupWrapper, StyledClose } from './styled'
+import { TransactionPopup } from './TransactionPopup'
 
 export function PopupItem({
   removeAfterMs,
@@ -39,11 +20,60 @@ export function PopupItem({
   content: PopupContent
   popKey: string
 }) {
+  const removePopup = useRemovePopup()
+  const removeThisPopup = useCallback(() => removePopup(popKey), [popKey, removePopup])
+  useEffect(() => {
+    if (removeAfterMs === null) return undefined
+
+    const timeout = setTimeout(() => {
+      removeThisPopup()
+    }, removeAfterMs)
+
+    return () => {
+      clearTimeout(timeout)
+    }
+  }, [removeAfterMs, removeThisPopup])
+
+  const theme = useContext(ThemeContext)
+
+  // mod
+  const isTxn = 'txn' in content
+  const isUnsupportedNetwork = 'unsupportedNetwork' in content
+  const isMetaTxn = 'metatxn' in content
+  const isWarningTxn = 'warning' in content
+
+  let popupContent
+  if (isTxn) {
+    const {
+      txn: { hash, success, summary },
+    } = content
+    popupContent = <TransactionPopup hash={hash} success={success} summary={summary} />
+  } else if (isMetaTxn) {
+    const {
+      metatxn: { id, success, summary },
+    } = content
+    popupContent = <TransactionPopup hash={id} success={success} summary={summary} />
+  } else if (isUnsupportedNetwork) {
+    popupContent = <FailedNetworkSwitchPopup chainId={content.failedSwitchNetwork} isUnsupportedNetwork />
+  } else if ('failedSwitchNetwork' in content) {
+    popupContent = <FailedNetworkSwitchPopup chainId={content.failedSwitchNetwork} />
+  } else if (isWarningTxn) {
+    popupContent = <WarningPopup warning={content.warning} />
+  }
+
+  const faderStyle = useSpring({
+    from: { width: '100%' },
+    to: { width: '0%' },
+    config: { duration: removeAfterMs ?? undefined },
+  })
+
   return (
-    <Wrapper>
-      <PopupItemWrapper removeAfterMs={removeAfterMs} content={content} popKey={popKey} />
-    </Wrapper>
+    <PopupItemWrapper>
+      <PopupWrapper css={content.styles}>
+        <StyledClose stroke={isUnsupportedNetwork ? theme.black : theme.text2} onClick={removeThisPopup} />
+        {popupContent}
+        {removeAfterMs !== null ? <AnimatedFader style={faderStyle} /> : null}
+      </PopupWrapper>
+    </PopupItemWrapper>
   )
 }
-
-export default PopupItem
